refactor: update primary bar icon and remove customization
Refs: SHELL-826
