+++ package.json
@@ -79,7 +79,7 @@
 	"dependencies": {
 		"@apollo/client": "3.9.10",
 		"@zextras/carbonio-design-system": "^7.1.0",
-		"@zextras/carbonio-shell-ui": ">=7.0.1-devel.1723474604916 <7.0.1",
+		"@zextras/carbonio-shell-ui": "file:zextras-carbonio-shell-ui-7.0.1.tgz",
 		"@zextras/carbonio-ui-preview": "^2.0.0",
 		"core-js": "^3.36.1",
 		"graphql": "^16.8.1",
